requirements.txt
Pillow==11.1.0
python-docx==1.1.2